Project/requirements.txt
asgiref==3.7.2
Django==4.1.7
djangorestframework==3.14.0
pip==23.2.1
pytz==2023.3.post1
setuptools==68.1.2
sqlparse==0.4.4
tzdata==2023.3
wheel==0.41.2
